requirements/base.txt
Django==1.8.3

django-allauth==0.19.1
django-suit==0.2.14
django-bootstrap3==5.4.0
django-recaptcha==1.0.4
django-getpaid==1.7.2
Pillow==2.9.0
django-money==0.7.4
